// src/app/api/model/category.ts
/**
 * Personal Finance Management API
 *
 * Contact: dragisa@yahoo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Category { 
    id?: number;
    name?: string;
    createdAt?: string;
    updatedAt?: string;
}