// src/app/api/model/recurringTransaction.ts
/**
 * Personal Finance Management API
 *
 * Contact: dragisa@yahoo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Transaction } from './transaction';


export interface RecurringTransaction { 
    id?: number;
    transaction?: Transaction;
    frequency?: string;
    nextOccurrence?: string;
    lastOccurrence?: string;
    status?: string;
    createdAt?: string;
    updatedAt?: string;
}